Check if linux before trying to install packages

+++ run.bash
@@ -37,14 +37,17 @@ copy_files() {
 }
 
 install_packages() {
-  $S apt-get update
-  $S apt-get install git vim 
-  for p in $(cat $CONFIG_FILES_DIR/PackagesList);
-  do
-    $S apt-get install -y $p
-  done
-  $S apt-get autoremove
-  $S apt-get autoclean
+  unamestr=`uname`
+  if [[ "$unamestr" == 'Linux' ]]; then
+    $S apt-get update
+    $S apt-get install git vim
+    for p in $(cat $CONFIG_FILES_DIR/PackagesList);
+    do
+      $S apt-get install -y $p
+    done
+    $S apt-get autoremove
+    $S apt-get autoclean
+  fi
 }
 
 configure() {
